removing unused  and refactor
