+++ src/types/mod.rs
@@ -12,7 +12,7 @@ mod tokenized;
 mod user;
 mod user_event;
 mod user_role;
-mod db;
+mod connection;
 pub use self::event::*;
 pub use self::login::*;
 pub use self::medium::*;
@@ -27,7 +27,7 @@ pub use self::tokenized::*;
 pub use self::user::*;
 pub use self::user_event::*;
 pub use self::user_role::*;
-pub use self::db::Connection;
+pub use self::connection::Connection;
 
 use juniper::Context as JuniperContext;
 use juniper::RootNode;
